Update form validation example
